Removed unnecessary code. Renamed Step.ts. PDA stepping lab works for multiple branches.
