+++ frontend/src/pages/NewFile/NewFile.js
@@ -126,7 +126,7 @@ const NewFile = () => {
         title="Turing Machine"
         description="Create a turing machine capable of representing recursively enumerable grammars."
         onClick={() => handleNewFile('TM')}
-        image={<TM />} CREATE A TM SVG
+        image={<TM />} 
       />
     </CardList>
